+++ src/helpers/transport/error.rs
@@ -28,7 +28,7 @@ pub enum Error {
 impl From<tokio_util::sync::PollSendError<TransportCommand>> for Error {
     fn from(source: tokio_util::sync::PollSendError<TransportCommand>) -> Self {
         let (command_name, query_id) = match source.into_inner() {
-            /// TODO: this is bad, requires matching for every command type
+            /// TODO: this is not optimal, requires matching for every command type
             Some(TransportCommand::NetworkEvent(data)) => {
                 (Some(NetworkEventData::name()), Some(data.query_id))
             }
